+++ pyoxidizer/src/pyrepackager/bytecodecompiler.py
@@ -7,6 +7,8 @@
 # When invoked, we start a server that listens for commands. We then
 # react to those commands and send results to the caller.
 
+import importlib._bootstrap_external
+import importlib.util
 import marshal
 import os
 import re
@@ -32,6 +34,7 @@ while True:
         name_len = stdin.readline().rstrip()
         source_len = stdin.readline().rstrip()
         optimize_level = stdin.readline().rstrip()
+        output_mode = stdin.readline().rstrip()
 
         name_len = int(name_len)
         source_len = int(source_len)
@@ -60,13 +63,33 @@ while True:
             encoding = 'utf-8'
             source = source[3:]
 
+        source_bytes = source
         source = source.decode(encoding)
 
         code = compile(source, name, 'exec', optimize=optimize_level)
         bytecode = marshal.dumps(code)
 
-        stdout.write(b'%d\n' % len(bytecode))
-        stdout.write(bytecode)
+        if output_mode == b'bytecode':
+            out = bytecode
+        elif output_mode == b'pyc-checked-hash':
+            source_hash = importlib.util.source_hash(source_bytes)
+            out = importlib._bootstrap_external._code_to_hash_pyc(
+                code,
+                source_hash,
+                checked=True,
+            )
+        elif output_mode == b'pyc-unchecked-hash':
+            source_hash = importlib.util.source_hash(source_bytes)
+            out = importlib._bootstrap_external._code_to_hash_pyc(
+                code,
+                source_hash,
+                checked=False,
+            )
+        else:
+            raise Exception('unknown output mode: %s' % output_mode)
+
+        stdout.write(b'%d\n' % len(out))
+        stdout.write(out)
         stdout.flush()
     else:
         raise Exception('invalid command: %s' % command)